+++ changelog.md
@@ -12,6 +12,7 @@
 - Fixed null derefs caused by Subtract Beam attempting to reuse stale health labels after switching floors
 - Fixed null deref caused by Library Cardtridge trying to read items from nonexistent shops when dropped
 - Fixed player being able to use Emergency Siren after firing beam weapons
+- Fixed Femtobyte being able to pick up used Gun Munchers
 
 ## 1.28.2 (2025-10-14)